+++ package.json
@@ -5,8 +5,7 @@
   "scripts": {
     "dev": "vite",
     "start": "vite",
-    "build": "vite build",
-    "build:safe": "tsc && vite build",
+    "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "repository": "https://github.com/riccox/meilisearch-ui",
@@ -18,52 +17,52 @@
   },
   "dependencies": {
     "@emotion/react": "^11.11.1",
-    "@hookform/resolvers": "^3.3.1",
-    "@mantine/core": "^7.0.0",
-    "@mantine/form": "^7.0.0",
-    "@mantine/hooks": "^7.0.0",
-    "@mantine/modals": "^7.0.0",
-    "@monaco-editor/react": "^4.5.2",
-    "@sira-ui/tailwind": "^0.11.6",
-    "@tabler/icons-react": "^2.34.0",
-    "@tanstack/react-query": "^4.35.3",
-    "@tanstack/react-query-devtools": "^4.35.3",
-    "ahooks": "^3.7.8",
-    "clsx": "^2.0.0",
-    "dayjs": "^1.11.10",
-    "echarts": "^5.4.3",
+    "@hookform/resolvers": "^3.1.1",
+    "@mantine/core": "^6.0.14",
+    "@mantine/form": "^6.0.14",
+    "@mantine/hooks": "^6.0.14",
+    "@mantine/modals": "^6.0.14",
+    "@monaco-editor/react": "^4.5.1",
+    "@sira-ui/tailwind": "^0.10.1",
+    "@tabler/icons-react": "^2.22.0",
+    "@tanstack/react-query": "^4.29.15",
+    "@tanstack/react-query-devtools": "^4.29.15",
+    "ahooks": "^3.7.7",
+    "clsx": "^1.2.1",
+    "dayjs": "^1.11.8",
+    "echarts": "^5.4.2",
     "echarts-for-react": "^3.0.2",
     "fuse.js": "^6.6.2",
     "immer": "^10.0.2",
     "lodash": "^4.17.21",
-    "meilisearch": "^0.34.2",
+    "meilisearch": "^0.33.0",
     "qs": "^6.11.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-error-boundary": "^4.0.11",
-    "react-hook-form": "^7.46.1",
+    "react-error-boundary": "^4.0.10",
+    "react-hook-form": "^7.45.0",
     "react-json-view": "^1.21.3",
-    "react-router-dom": "^6.16.0",
-    "sonner": "^0.7.4",
-    "zod": "^3.22.2",
-    "zustand": "^4.4.1"
+    "react-router-dom": "^6.13.0",
+    "sonner": "^0.5.0",
+    "zod": "^3.21.4",
+    "zustand": "^4.3.8"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.198",
-    "@types/node": "^20.6.3",
-    "@types/qs": "^6.9.8",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
-    "@vitejs/plugin-react": "^4.0.4",
-    "autoprefixer": "^10.4.15",
-    "eslint": "8.47.0",
+    "@types/lodash": "^4.14.195",
+    "@types/node": "^20.3.1",
+    "@types/qs": "^6.9.7",
+    "@types/react": "^18.2.13",
+    "@types/react-dom": "^18.2.6",
+    "@vitejs/plugin-react": "^4.0.1",
+    "autoprefixer": "^10.4.14",
+    "eslint": "8.43.0",
     "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "postcss": "^8.4.30",
-    "prettier": "^3.0.3",
-    "tailwindcss": "^3.3.3",
-    "typescript": "5.1.6",
-    "vite": "^4.4.9"
+    "postcss": "^8.4.24",
+    "prettier": "^2.8.8",
+    "tailwindcss": "^3.3.2",
+    "typescript": "5.0.4",
+    "vite": "^4.3.9"
   },
   "version": "0.4.1"
 }
