+++ src/components/Header/index.tsx
@@ -1,6 +1,5 @@
 import { ActionIcon, HoverCard } from '@mantine/core';
 import {
-  IconAffiliate,
   IconArrowsLeftRight,
   IconBook2,
   IconBooks,
@@ -146,7 +145,6 @@ export const Header: FC<Props> = ({ client, className }) => {
               <IconBooks size={14} />
               <p>Index</p>
             </Link>
-
             <div
               onClick={() => {
                 navigate([`/ins/${currentInstance.id}/keys`], { currentInstance });
