+++ src/pages/dashboard.tsx
@@ -186,7 +186,7 @@ function Dashboard() {
             <Tooltip position={'bottom'} label="Keys">
               <ActionIcon
                 variant="light"
-                color="grape"
+                color="purple"
                 onClick={() => onClickInstance(instance, `/ins/${instance.id}/keys`)}
               >
                 <IconKey size={24} />
@@ -200,7 +200,7 @@ function Dashboard() {
 
   return (
     <div className="bg-mount full-page justify-center items-center gap-y-6">
-      <div className={`w-1/2 2xl:w-1/4 h-2/3 flex flex-col justify-center items-center gap-y-10`}>
+      <div className={`w-1/4 h-2/3 flex flex-col justify-center items-center gap-y-10`}>
         <Logo />
         <p className={`text-brand-2 font-bold xl:text-3xl text-xl w-screen text-center`}>A Beautiful Meilisearch UI</p>
         <div className={`grid grid-cols-1 gap-y-3 w-full p-1  overflow-y-scroll`}>
@@ -246,6 +246,7 @@ function Dashboard() {
           <Tooltip
             position={'bottom-start'}
             multiline
+            width={400}
             label={`Remember enable CORS in your instance server for this ui domain first`}
           >
             <Autocomplete
@@ -260,6 +261,7 @@ function Dashboard() {
           <Tooltip
             position={'bottom-start'}
             multiline
+            width={400}
             label="Don't care! Your instance config will only be store in your local browser"
           >
             <PasswordInput
