+++ src/pages/task/index.tsx
@@ -144,6 +144,7 @@ function Tasks() {
 
             <Select
               placeholder="Filter Task Type"
+              clearable
               radius={'lg'}
               data={[
                 { value: TaskTypes.DOCUMENT_DELETION, label: 'DOCUMENT_DELETION' },
@@ -157,6 +158,7 @@ function Tasks() {
             />
             <Select
               placeholder="Filter Task Status"
+              clearable
               radius={'lg'}
               data={[
                 { value: 'succeeded', label: 'Succeeded ✅' },
@@ -197,10 +199,9 @@ function Tasks() {
           opened={isTaskDetailModalOpen}
           onClose={() => setIsTaskDetailModalOpen(false)}
           title="Task Detail"
-          size={'xl'}
         >
           <div className="flex justify-center items-center">
-            <Code block className="w-full">
+            <Code block className="w-96">
               {stringifyJsonPretty(taskDetailModalContent)}
             </Code>
           </div>
