+++ src/providers/Mantine.tsx
@@ -1,9 +1,19 @@
 import type { FC, ReactNode } from 'react';
-import { MantineProvider } from '@mantine/core';
+import { createEmotionCache, DefaultMantineColor, MantineProvider, Tuple } from '@mantine/core';
 import theme from '@/src/style/theme.json';
 import _ from 'lodash';
 import { ModalsProvider } from '@mantine/modals';
-import '@mantine/core/styles.css';
+
+type CustomColors = 'brand';
+type ExtendedCustomColors = CustomColors | DefaultMantineColor;
+
+declare module '@mantine/core' {
+  export interface MantineThemeColorsOverride {
+    colors: Record<ExtendedCustomColors, Tuple<string, 10>>;
+  }
+}
+
+const myCache = createEmotionCache({ key: 'mantine' });
 
 type Props = {
   children: ReactNode;
@@ -12,9 +22,10 @@ type Props = {
 export const MantineUIProvider: FC<Props> = ({ children }) => {
   return (
     <MantineProvider
+      emotionCache={myCache}
+      withNormalizeCSS
       theme={{
-        // @ts-ignore
-        colors: _.pick(theme.colors, 'brand'),
+        colors: _.pick(theme.colors as unknown as Record<CustomColors, Tuple<string, 10>>, 'brand'),
         primaryColor: 'brand',
       }}
     >
