+++ src/style/theme.json
@@ -6,16 +6,16 @@
       "light": "#fafafa"
     },
     "brand": [
-      "#ffe7f7",
-      "#ffcee7",
-      "#ff9bcb",
-      "#ff64ad",
-      "#fe3694",
-      "#fe1984",
-      "#ff027c",
-      "#e4006a",
-      "#cd005e",
-      "#b40051"
+      "#ffe2f5",
+      "#ffb1d8",
+      "#ff7fbd",
+      "#ff4da2",
+      "#fe1e87",
+      "#e5076e",
+      "#b30055",
+      "#81003d",
+      "#4f0025",
+      "#1f000e"
     ]
   }
 }
